requirements.txt
# LangChain ecosystem
langchain==0.3.0
langchain-community
langchain-core
langchain-text-splitters
langchain-openai
langchain-huggingface
langchain-chroma
langchain-groq
# LangServe (required for serving chains)
langserve==0.3.1
fastapi==0.111.1
pydantic==2.9.2
uvicorn==0.30.6
python-dotenv
ipykernel
pypdf
bs4
arxiv
pymupdf
wikipedia
chromadb
sentence_transformers
faiss-cpu
duckdb
pandas
openai
duckduckgo_search==5.3.1b1
mysql-connector-python
SQLAlchemy
validators==0.28.1
youtube_transcript_api
unstructured
pytube
numexpr
huggingface_hub
hf_xet
streamlit
pymongo>=4.3.0
typing-extensions>=4.7.1



